+++ app/src/main/java/com/amory/departmentstore/model/SanPham.kt
@@ -6,7 +6,7 @@ import java.io.Serializable
 data class SanPham(
     val id: Int,
     val name: String,
-    val price: Float,
+    val price: Double,
     val imageUrl: String,
     val description: String,
     val createdAt: String,
